pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "short-term-trading"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "yfinance>=0.2.31",
    "pyarrow>=15.0.0",
    "mplfinance>=0.12.10b0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = ["jupyter", "ipykernel"]

[tool.setuptools.packages.find]
include = ["core*", "strategies*", "backtest*", "charts*"]
